Moved actyx-functions to separate file; Code cleanup;
